creacion y llenado de tabla estudiantes

+++ Models/Student.cs
@@ -4,7 +4,7 @@
     {
         public string name { get; set; }
         public int age { get; set; }
-        public int document { get; set; }
+        public long document { get; set; }
         public string licence { get; set; }
     }
 }
